+++ views/layouts/dashboard.php
@@ -190,9 +190,9 @@ JS
                             'url' => Url::to(['/admin/admin/index'])
                         ],
                         'links' => isset($this->params['breadcrumbs']) ? $this->params['breadcrumbs'] : [],
-                    ]) ?>
-                    <?= Alert::widget() ?>
-                    <?= $content ?>
+                    ]); ?>
+                    <?= Alert::widget(); ?>
+                    <?= $content; ?>
                     <?php Pjax::end(); ?>
                 </div>
             </div>